mvp_backend/parser/requirements.txt
python-docx>=1.1.2
pdfplumber>=0.11.4
pytesseract>=0.3.13
Pillow>=10.4.0
pymupdf>=1.24.10
pdf2image>=1.17.0
openpyxl>=3.1.5
